fix(multitask): keep orchestrator bricks under line limit
Copilot-Session: 8b10b81e-2dc0-4afb-9673-d9d1c5805637
